coupon ui redesign, no validation yet

+++ src/main/java/com/skiipr/server/controller/dashboard/SettingsController.java
@@ -14,6 +14,7 @@ import com.skiipr.server.model.DAO.CouponDao;
 import com.skiipr.server.model.DAO.MerchantDao;
 import com.skiipr.server.model.DAO.PlanDao;
 import com.skiipr.server.model.Merchant;
+import com.skiipr.server.model.form.CouponForm;
 import com.skiipr.server.model.form.SettingsForm;
 import com.skiipr.server.model.form.PaymentOptions;
 import com.skiipr.server.model.validators.MerchantValidator;
@@ -86,39 +87,35 @@ public class SettingsController {
         return "/dashboard/settings";
     }
     @RequestMapping(value = "/dashboard/settings/discountcodes", method = RequestMethod.GET)
-    public String viewDiscountCodes(ModelMap model, HttpServletRequest httpServletRequest){
-        if(httpServletRequest.getSession().getAttribute("flash") != null){
-            FlashNotification flash = (FlashNotification) httpServletRequest.getSession().getAttribute("flash");
-            model.addAttribute("flash", flash);
-            httpServletRequest.getSession().removeAttribute("flash");
-        }
-        List<Coupon> coupons = couponDao.findAllByMerchant();
-        model.addAttribute("coupons", coupons);
-        return "/dashboard/settings/discountcodes";
-        
-    }
-    
-    @RequestMapping(value = "/dashboard/settings/discountcodes/new", method = RequestMethod.POST)
-    public String createDiscountCode(@Valid Coupon coupon, BindingResult bindingResult, Model uiModel, HttpServletRequest httpServletRequest) {
-        if (bindingResult.hasErrors()) {
-            uiModel.addAttribute("coupon", coupon);
-            return "/dashboard/settings/discountcodes/create";
-        }
-        uiModel.asMap().clear();
-        coupon.setMerchantID(sessionUser.getUser().getMerchantId());
-        couponDao.save(coupon);
-        uiModel.addAttribute("flash", FlashNotification.create(Status.SUCCESS, "Product Added"));
-        return show(coupon.getCouponID(), uiModel);
-    }
-    
-     @RequestMapping(value = "/dashboard/settings/discountcodes/new", method = RequestMethod.GET)
-    public String createForm(Model uiModel) {
-        uiModel.addAttribute("coupon", new Coupon());
+    public String listCoupons(@RequestParam(value = "page", required = false) Integer page, @RequestParam(value = "size", required = false) Integer size, ModelMap modelMap) {
+        Integer sizeNo = this.getCouponPageSize(size);
+        Integer startPage = this.getCouponStartPage(page, sizeNo);
         List<CouponType> couponTypes = new ArrayList<CouponType>(Arrays.asList(CouponType.values()));
-        uiModel.addAttribute("couponTypes", couponTypes);
-        
-        
-        return "/dashboard/settings/discountcodes/create";
+        modelMap.addAttribute("couponTypes", couponTypes);
+        modelMap.addAttribute("coupons", couponDao.findRange(startPage, sizeNo));
+        modelMap.addAttribute("couponModel", new Coupon());
+        modelMap.addAttribute("maxPages", this.getCouponMaxPages(sizeNo));
+        return "/dashboard/settings/discountcodes";
+    }
+    
+    @RequestMapping(value = "/dashboard/settings/discountcodes", method = RequestMethod.PUT)
+    public String create(CouponForm formCoupon, BindingResult bindingResult, ModelMap modelMap) {
+       formCoupon.setCouponID(0l);
+       if (!formCoupon.validate(couponDao, bindingResult)) {
+            modelMap.addAttribute("flash", FlashNotification.create(Status.FAILURE, "There was an error creating your coupon"));
+            
+            modelMap.addAttribute("openProdID", 0);
+       }else{
+            Coupon coupon = new Coupon();
+            formCoupon.setAttributes(coupon);
+            
+            
+            coupon.setMerchantID(sessionUser.getUser().getMerchantId());
+            couponDao.save(coupon);
+            modelMap.addAttribute("openCoupID", coupon.getCouponID());
+            modelMap.addAttribute("flash", FlashNotification.create(Status.SUCCESS, "Coupon Created"));
+        }
+        return listCoupons(null, null, modelMap);
     }
      
     @RequestMapping(value = "/dashboard/settings/discountcodes/view/{id}", method = RequestMethod.GET)
@@ -128,11 +125,16 @@ public class SettingsController {
         return "/dashboard/settings/discountcodes/view";
     }
     
-    @RequestMapping(value = "/dashboard/settings/discountcodes/delete/{id}", method = RequestMethod.GET)
-    public String deleteCoupon(@PathVariable("id") Long id, @RequestParam(value = "page", required = false) Integer page, @RequestParam(value = "size", required = false) Integer size, Model uiModel) {
+    @RequestMapping(value = "/dashboard/settings/discountcodes", method = RequestMethod.DELETE)
+    public String deleteCoupon(@RequestParam("couponID") Long id, ModelMap model, HttpServletRequest httpServletRequest) {
         Coupon coupon = couponDao.findByIDandMerchant(id);
-        couponDao.delete(coupon);
-        return "redirect:/dashboard/settings/discountcodes";
+        if(coupon == null){
+           model.addAttribute("flash", FlashNotification.create(Status.FAILURE, "This Coupon does not belong to you or is invalid."));    
+        }else{
+            couponDao.delete(coupon);
+            model.addAttribute("flash", FlashNotification.create(Status.SUCCESS, "Coupon deleted"));
+        }
+        return listCoupons(null, null, model);
         
     }
     
@@ -178,28 +180,26 @@ public class SettingsController {
         return viewSecurity(model);
     }
     
-    @RequestMapping(value = "/dashboard/settings/discountcodes/edit", method = RequestMethod.PUT)
-    public String updateCoupon(@Valid Coupon coupon, BindingResult bindingResult, Model uiModel, HttpServletRequest httpServletRequest) {
-        if (bindingResult.hasErrors()) {
-            uiModel.addAttribute("coupon", coupon);
-            return updateCouponForm(coupon.getCouponID(), uiModel);
-        }
-        uiModel.asMap().clear();
-        uiModel.addAttribute("flash", FlashNotification.create(Status.SUCCESS, "Product Updated"));
-        coupon.setMerchantID(sessionUser.getUser().getMerchantId());
-        couponDao.update(coupon);
-        return updateCouponForm(coupon.getCouponID(), uiModel);
-    }
-    
-    @RequestMapping(value = "/dashboard/settings/discountcodes/edit/{id}", method = RequestMethod.GET)
-    public String updateCouponForm(@PathVariable("id") Long id, Model uiModel) {
-        Coupon coupon = couponDao.findByIDandMerchant(id);
-        if (coupon == null){
-            return "redirect:/dashboard/settings/discountcodes";
+    @RequestMapping(value = "/dashboard/settings/discountcodes", method = RequestMethod.POST)
+     public String updateCoupon(@RequestParam(value = "page", required = false) Integer page, @RequestParam(value = "size", required = false) Integer size, CouponForm formCoupon, BindingResult bindingResult, ModelMap modelMap) {
+       if (!formCoupon.validate(couponDao, bindingResult)) {
+            modelMap.addAttribute("flash", FlashNotification.create(Status.FAILURE, "There was an error updating your coupon"));
+            return listCoupons(page, size, modelMap);
         }else{
-            uiModel.addAttribute("coupon", coupon);
-            return "/dashboard/settings/discountcodes/update";  
+            Coupon coupon = couponDao.findByIDandMerchant(formCoupon.getCouponID());
+            if(coupon == null){
+                modelMap.addAttribute("flash", FlashNotification.create(Status.FAILURE, "This coupon does not exist."));
+            }else{
+                
+                formCoupon.setAttributes(coupon);
+                
+                couponDao.update(coupon);
+                
+                modelMap.addAttribute("flash", FlashNotification.create(Status.SUCCESS, "Coupon Updated"));
+            }
         }
+        modelMap.addAttribute("openCoupID", formCoupon.getCouponID());
+        return listCoupons(page, size, modelMap);
     }
     
     @RequestMapping(value = "/dashboard/settings/paymentoptions", method = RequestMethod.POST)
@@ -249,5 +249,30 @@ public class SettingsController {
         return "/dashboard/settings/paymentoptions";
     }
     
-   
+    private Integer getCouponPageSize(Integer size){
+        if(size == null){
+            return 10;
+        }else{
+            return size;
+        }
+    }
+    
+    private Integer getCouponStartPage(Integer page, Integer pageSize){
+        if(page == null){
+            return 0;
+        }else{
+            return (page - 1) * pageSize;
+        }
+    }
+    
+    private Integer getCouponMaxPages(Integer sizeNo){
+        float nrOfPages = (float) couponDao.countByMerchant() / sizeNo;
+        if(nrOfPages > (int) nrOfPages || nrOfPages == 0.0){
+            return (int) nrOfPages + 1;
+        }else{
+            return (int) nrOfPages;
+        }
+    }
 }
+    
+
